+++ grako/parsing.py
@@ -1,20 +1,6 @@
 # -*- coding: utf-8 -*-
 # Copyright (C) 2017      by Juancarlo Añez
 # Copyright (C) 2012-2016 by Juancarlo Añez and Thomas Bragg
-"""
-Parser is the base class for generated parsers and for the bootstrap parser
-(the parser that parses Grako grammars).
-
-Parser does memoization at the rule invocation level, and provides the
-decorators, context managers, and iterators needed to make generated parsers
-simple.
-
-Parser is also in charge of dealing with comments, with the help of
-the .buffering module.
-
-Parser.parse() will take the text to parse directly, or an instance of the
-.buffeing.Buffer class.
-"""
 from __future__ import absolute_import, division, print_function, unicode_literals
 
 from grako.exceptions import FailedRef