+++ CHANGELOG.md
@@ -27,9 +27,12 @@ The format of this *Change Log* is inspired by [keeapachangelog.org].
 - [PyPi] (the _Python Package Index_) now requires there be a single source-code archive in each release. The chosen format is '.zip'.
 - [PyPi] publication requires there be a `README`, `README.rst`, or `README.txt` file. Renamed `DISTRIBUTION.rst` to `README.rst`.
 - Make `parseinfo` less special so it can be part of `ATS` and of `asjson(ast)`.
-- Move `etc/grako.ebnf`, the **Grako** grammar to a `grammar/` directory.
+- Move `etc/grako.ebnf`, the **Grako** grammar, to a `grammar/` directory.
 - Deprecated the `?/regexp/?` style of patterns.
 - Deprecated the `(* *)` style of comments.
+- Renamed classes so the entity names are meaningful instead of just being references to the tool's name (`EBNFParser` instead of `GrakoParser`, `ParseException` instead of
+  `GrakoException`, ...).
+- Removed classes that had become irrelevant (`GrakoContext`, `GrakoParserBase`, ...)
 - Updated copyright notices.